# specs/001-project-kensh-poc/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Project Kenshō POC API
  description: All-Rust anime streaming platform API specification
  version: 0.1.0
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.kensho.app/api
    description: Production server
security:
  - BearerAuth: []
paths:
  /anime/{id}:
    get:
      summary: Get anime details
      description: Retrieve comprehensive metadata for a specific anime including relationships
      operationId: getAnimeById
      tags:
        - Anime
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Anime details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimeDetail'
        '404':
          description: Anime not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AnimeSummary:
      type: object
      required:
        - id
        - title
        - poster_url
        - episodes
        - status
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        poster_url:
          type: string
          format: uri
        episodes:
          type: integer
        status:
          type: string
          enum: [finished, ongoing, upcoming, unknown]
        anime_type:
          type: string
          enum: [tv, movie, ova, ona, special, unknown]
        imdb_rating:
          type: number
          format: float
          minimum: 0
          maximum: 10
    AnimeDetail:
      allOf:
        - $ref: '#/components/schemas/AnimeSummary'
        - type: object
          properties:
            synonyms:
              type: array
              items:
                type: string
            synopsis:
              type: string
            anime_season:
              type: object
              properties:
                season:
                  type: string
                  enum: [spring, summer, fall, winter]
                year:
                  type: integer
            imdb:
              type: object
              properties:
                id:
                  type: string
                rating:
                  type: number
                  format: float
                votes:
                  type: integer
            tags:
              type: array
              items:
                $ref: '#/components/schemas/Tag'
            related_anime:
              type: object
              properties:
                sequels:
                  type: array
                  items:
                    $ref: '#/components/schemas/AnimeSummary'
                prequels:
                  type: array
                  items:
                    $ref: '#/components/schemas/AnimeSummary'
                related:
                  type: array
                  items:
                    $ref: '#/components/schemas/AnimeSummary'
...
    Tag:
      type: object
      required:
        - id
        - name
        - category
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
          enum: [genre, theme, demographic, content]
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object